blucon.sh: Allow validation/results dir override
When running in CI, the validation git repository is cloned in the
job workspace, which usually has a custom directory name matching the
job name (e.g. bluval-daily-master) instead of the previously hardcoded
'validation' directory name.
Also, the results dir should not be a sibling of the validation git repo
directory when running in CI, but a subdirectory of the said dir.

To accomodate this, allow overriding both previously hardcoded directory
paths via environment variables (VALIDATION_DIR, RESULTS_DIR).

While at it, add a new '-t' argument for specifying the Docker tag to be
used for the validation docker images (previously hardcoded to 'latest')
in order to support running docker images associated to a specific
Validation release tag (e.g. 3.0.0).

JIRA: VAL-95

Change-Id: I4d3c10881de28c64bcca05ff23aa1025f67a1f5e
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>